\name{supern-package}
\alias{supern-package}
\alias{supern}
\docType{package}
\title{
\packageTitle{supern}
}
\description{
\packageDescription{supern}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{supern}
\packageIndices{supern}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{supern}

Maintainer: \packageMaintainer{supern}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
